+++ src/headless/shared/index.js
@@ -1,0 +1,7 @@
+import * as parsers from './parsers.js';
+import * as constants from './constants.js';
+import api from './api/index.js';
+import _converse from './_converse';
+import i18n from './i18n';
+
+export { _converse, api, constants, i18n, parsers };